+++ codex-rs/core/src/tools/handlers/apply_patch.rs
@@ -1,15 +1,42 @@
 use std::collections::HashMap;
 
-use async_trait::async_trait;
-
+use crate::apply_patch;
+use crate::apply_patch::ApplyPatchExec;
+use crate::apply_patch::InternalApplyPatchInvocation;
+use crate::apply_patch::convert_apply_patch_to_protocol;
+use crate::codex::ApplyPatchCommandContext;
+use crate::codex::ExecCommandContext;
+use crate::codex::Session;
+use crate::codex::TurnContext;
+use crate::error::CodexErr;
+use crate::error::SandboxErr;
 use crate::exec::ExecParams;
+use crate::exec::ExecToolCallOutput;
+use crate::exec::StdoutStream;
+use crate::executor::ExecutionMode;
+use crate::executor::errors::ExecError;
+use crate::executor::linkers::PreparedExec;
 use crate::function_tool::FunctionCallError;
+use crate::tools::{handle_container_exec_with_params, MODEL_FORMAT_HEAD_BYTES};
+use crate::tools::MODEL_FORMAT_HEAD_LINES;
+use crate::tools::MODEL_FORMAT_MAX_BYTES;
+use crate::tools::MODEL_FORMAT_MAX_LINES;
+use crate::tools::MODEL_FORMAT_TAIL_LINES;
 use crate::tools::context::ToolInvocation;
 use crate::tools::context::ToolOutput;
 use crate::tools::context::ToolPayload;
 use crate::tools::registry::ToolHandler;
 use crate::tools::registry::ToolKind;
 use crate::tools::spec::ApplyPatchToolArgs;
+use crate::turn_diff_tracker::TurnDiffTracker;
+use async_trait::async_trait;
+use codex_apply_patch::MaybeApplyPatchVerified;
+use codex_apply_patch::maybe_parse_apply_patch_verified;
+use codex_protocol::protocol::AskForApproval;
+use codex_utils_string::take_bytes_at_char_boundary;
+use codex_utils_string::take_last_bytes_at_char_boundary;
+use serde::Serialize;
+use tracing::trace;
 
 pub struct ApplyPatchHandler;
 
@@ -66,7 +93,7 @@ impl ToolHandler for ApplyPatchHandler {
             justification: None,
         };
 
-        let content = crate::codex::handle_container_exec_with_params(
+        let content = handle_container_exec_with_params(
             tool_name.as_str(),
             exec_params,
             session,
