+++ codex-rs/core/src/tools/handlers/shell.rs
@@ -8,6 +8,7 @@ use crate::function_tool::FunctionCallError;
 use crate::tools::context::ToolInvocation;
 use crate::tools::context::ToolOutput;
 use crate::tools::context::ToolPayload;
+use crate::tools::handle_container_exec_with_params;
 use crate::tools::registry::ToolHandler;
 use crate::tools::registry::ToolKind;
 
@@ -62,7 +63,7 @@ impl ToolHandler for ShellHandler {
                         ))
                     })?;
                 let exec_params = Self::to_exec_params(params, turn);
-                let content = crate::codex::handle_container_exec_with_params(
+                let content = handle_container_exec_with_params(
                     tool_name.as_str(),
                     exec_params,
                     session,
@@ -79,7 +80,7 @@ impl ToolHandler for ShellHandler {
             }
             ToolPayload::LocalShell { params } => {
                 let exec_params = Self::to_exec_params(params, turn);
-                let content = crate::codex::handle_container_exec_with_params(
+                let content = handle_container_exec_with_params(
                     tool_name.as_str(),
                     exec_params,
                     session,
